requirements.txt
alembic==1.4.2
autoenv==1.0.0
blinker==1.4
click==8.1.3
dominate==2.7.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.40
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
